% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{colony_similarity}
\alias{colony_similarity}
\title{Calculate similarity between two colonies/sites}
\usage{
colony_similarity(data, colony1, colony2)
}
\arguments{
\item{data}{A data frame containing seabird counts}

\item{colony1}{First colony name}

\item{colony2}{Second colony name}
}
\value{
A numeric value representing similarity (0 = different, 1 = identical)
}
\description{
Computes Bray-Curtis similarity index between two sites based on species counts
}
\examples{
colony_similarity(seabird_counts, "Foula", "Fair Isle")

}
