% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{list_colonies}
\alias{list_colonies}
\title{List available colonies/sites}
\usage{
list_colonies(data)
}
\arguments{
\item{data}{A data frame containing seabird counts}
}
\value{
A character vector of colony names
}
\description{
Returns a vector of unique site names in the dataset subset.
}
\examples{
list_colonies(seabird_counts)

}
