% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_abundance.R
\name{relative_abundance}
\alias{relative_abundance}
\title{Calculate relative abundance of species in a colony/site.}
\usage{
relative_abundance(data, colony)
}
\arguments{
\item{data}{A data frame containing seabird counts}

\item{colony}{Name of the colony/site}
}
\value{
A data frame with species and their relative abundance
}
\description{
Computes the proportion of each species within a given site.
}
\examples{
relative_abundance(seabird_counts, "Foula")

}
