% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seabird_counts}
\alias{seabird_counts}
\title{Seabird colony count data}
\format{
A data frame with:
\describe{
\item{colony}{Colony/site name}
\item{species}{Seabird species name}
\item{count}{Count of individuals}
}
}
\source{
Joint Nature Conservation Committee (JNCC).
Seabirds Count dataset.
\url{https://jncc.gov.uk/our-work/seabirds-count/}
}
\usage{
seabird_counts
}
\description{
A curated subset of seabird colony counts derived from the JNCC
Seabirds Count dataset.
}
\details{
The dataset has been simplified fo demonstration and teaching purposes.
}
\keyword{datasets}
