% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness.R
\name{species_richness}
\alias{species_richness}
\title{Calculate species richness for a colony/site.}
\usage{
species_richness(data, colony)
}
\arguments{
\item{data}{A data frame containing seabird counts}

\item{colony}{Name of the colony/site}
}
\value{
An integer giving the number of species present.
}
\description{
Counts the number of unique seabird species observed in a given colony.
}
\examples{
species_richness(seabird_counts, "Foula")

}
